refactor(python): move EventConverter and PartConverter to modular a2a subpackage
Extracts A2uiEventConverter and A2uiPartConverter from the bloated
send_a2ui_to_client_toolset.py module and moves them to separate source
and test modules under the adk/a2a package.
