+++ samples/agent/adk/mcp_app_proxy/agent_executor.py
@@ -18,7 +18,7 @@ from typing import override
 from a2a.server.agent_execution import RequestContext
 from a2a.types import AgentCapabilities, AgentCard, AgentExtension, AgentSkill
 from a2ui.a2a.extension import try_activate_a2ui_extension
-from a2ui.adk.send_a2ui_to_client_toolset import A2uiEventConverter
+from a2ui.adk.a2a.event_converter import A2uiEventConverter
 from a2ui.schema.constants import A2UI_CLIENT_CAPABILITIES_KEY
 from a2ui.schema.manager import A2uiSchemaManager
 from google.adk.a2a.converters.request_converter import AgentRunRequest
